🚚 move error fallback for dep-cruiser
